+++ src/main/java/sg/edu/ntu/singastays/repositories/AttractionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import sg.edu.ntu.singastays.entities.Attraction;
 
-public interface AttractionRepository extends JpaRepository<Attraction, Integer> {
+public interface AttractionRepository extends JpaRepository<Attraction, Long> {
 
 }
